change type of smLength in SmppSubmitSmBody from byte to short, to support message length 0-254 (#15)

+++ netty-codec-smpp/src/main/java/io/github/protocol/codec/smpp/SmppDecoder.java
@@ -244,7 +244,7 @@ public class SmppDecoder extends LengthFieldBasedFrameDecoder {
         byte replaceIfPresentFlag = frame.readByte();
         byte dataCoding = frame.readByte();
         byte smDefaultMsgId = frame.readByte();
-        byte smLength = frame.readByte();
+        short smLength = frame.readShort();
         byte[] shortMessage = new byte[smLength];
         frame.readBytes(shortMessage);
         return new SmppSubmitSmBody(serviceType, sourceAddrTon, sourceAddrNpi, sourceAddr,
